create onSubmit handler funcion in a different file to make code clear
