+++ src/handlers/index.js
@@ -1,0 +1,3 @@
+export { onClickHandlerForEmailFunction } from "./onClickHandlerForEmailFunction";
+export { onClickHandlerForPasswordFunction } from "./onClickHandlerForPasswordFunction";
+export { onSubmitHandlerForLogin } from "./onSubmitHandlerForLogin";
